Filled in the note a bit more.  I also give the details of an alternate presentation of FILL that I think it quite elegant, and makes the cut-elimination procedure very elegant.
There is one problem with some of the transformations given in the
section on cut-elimination which is that in some instances the right
side of some sequents is allowed to be empty, but my transformations are
excluding this possibility.  I am going to fix this next.  I am hoping to
have this part done by Monday.