Major Layout Changes
I didn't realize it until now, but we were using \usepackage{fullpage}
which substantially decreased the size of the margins.  The submission
version no longer uses it, but I kept it in the report, because the
large derivations in the cut-elimination proof will become unreadable.

The submission version went from 14 pages to 19 pages.  To shrink down
the page length I moved some proofs into the report.  We are now at 16
pages.